+++ CHANGELOG.md
@@ -188,6 +188,8 @@ TODO
 - Added ScionDatagram packet
   [#278](https://github.com/scionproto-contrib/jpan/pull/278)
   [#279](https://github.com/scionproto-contrib/jpan/pull/279)
+- Closeable PathSelectors and close() calling SimpleCache
+  [#280](https://github.com/scionproto-contrib/jpan/pull/280)
 
 ### Fixed